+++ project-base/CHANGELOG.md
@@ -50,6 +50,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - services instanced after DI container creation are set as synthetic
 - users and administrators are logged out of all the sessions except the current one on password change (this is required in Symfony 4) (@TomasLudvik)
 - running Phing without parameter now shows list of available targets instead of building application (@TomasLudvik)
+- updated presta/sitemap-bundle to version 1.5.2 in order to avoid deprecated calls (@TomasLudvik)
+ - updated SitemapListener to avoid using of deprecated SitemapListenerInterface
+ 
 
 ### Fixed
 - `BrandFacade::create()` now generates friendly URL for all domains (@sspooky13)